backend/app/requirements.txt
streamlit
PyPDF2
sentence-transformers
chromadb
Pillow
pdf2image
torch
transformers
pytesseract
protobuf==3.20.3